+++ CHANGELOG.md
@@ -39,12 +39,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - User input page formatted to clearly separate different categories and selectors by topic
   - Message/File pane more compact, cleaner, making it easier to browse many categories
   - Messages and Files more streamlined and less 'Chunky' easier to browse multiple files/messages
+  - Reduced 'over-scrolling'
+- loops (parallel runs) available for all workflows instead of having it's own specific workflow.
 
 ### Removed
 
 - o1-preview (o1 is released but not as good as o3-mini, while being *much* more expensive)
 - staging-directory: was used at the end of a request, saving files (typically in threads) is faster.
 - EncyclopediaManagement - made redundant by internet search
+- Loops workflow
 
 ## [0.9.5] - 2025-02-18 - Pre public beta